requirements.txt
# -----------------------------
# Core Flask Framework
# -----------------------------
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.1.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==2.1.5
email_validator==2.2.0

# -----------------------------
# Data / Math / Utilities
# -----------------------------
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
python-dateutil==2.9.0.post0
pytz==2024.1
packaging==24.1

# -----------------------------
# Firebase / Google Cloud
# -----------------------------
firebase-admin==6.5.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
googleapis-common-protos==1.63.2
grpcio==1.66.1
protobuf==4.25.3

# ---- Pyrebase4 (legacy but working) ----
pyrebase4==4.7.1
requests-toolbelt==0.10.1
PyJWT==2.9.0
rsa==4.9
cachetools==5.5.0

# -----------------------------
# Web / Requests stack
# -----------------------------
requests==2.29.0              # ✅ fits both pyrebase4 (<2.30) and google-api-core (>=2.18)
urllib3==1.26.19              # ✅ required for appengine adapter
idna==3.7
certifi==2024.8.30
chardet==5.2.0

# -----------------------------
# HTML / Parsing / Markdown
# -----------------------------
PyYAML==6.0.2
click==8.1.7
colorama==0.4.6
termcolor==2.4.0
blinker==1.8.2
visitor==0.1.3
dominate==2.9.1
soupsieve==2.5
beautifulsoup4==4.12.3
markdown==3.6

# -----------------------------
# Testing
# -----------------------------
pytest==8.3.3
attrs==24.2.0
pluggy==1.5.0
atomicwrites==1.4.1
more-itertools==10.3.0
wcwidth==0.2.13
